refactor: CardElement 및 UnityTextStyle 모델 구조 개선 (#27)
- `CardElementWithAssetKeys`를 DTO로 명시하여 역할 분리 및 명확화
- Transform 속성을 Non-null로 변경하고 기본값을 설정하여 데이터 안정성 확보
- `UnityTextStyle`의 미사용 속성 제거 및 기본 설정 최적화
- Mapper 내 미사용 텍스트 속성 제거
