+++ core/data/src/main/java/com/itschristmas/data/mapper/CardElementWithAssetKeysMapper.kt
@@ -1,9 +1,10 @@
 package com.itschristmas.data.mapper
 
+import com.itschristmas.database.dto.CardElementWithAssetKeysDto
 import com.itschristmas.domain.model.CardElementWithAssetKeys
-import com.itschristmas.database.entity.CardElementWithAssetKeysEntity
 
-fun CardElementWithAssetKeysEntity.toDomain(): CardElementWithAssetKeys {
+
+fun CardElementWithAssetKeysDto.toDomain(): CardElementWithAssetKeys {
     return CardElementWithAssetKeys(
         cardElement = cardElement.toDomain(),
         unityKey = unityKey,
