+++ core/domain/src/main/java/com/itschristmas/domain/model/UnityTextStyle.kt
@@ -4,12 +4,10 @@ import com.itschristmas.domain.enum.UnityFontWeight
 import com.itschristmas.domain.enum.UnityTextAlign
 
 data class UnityTextStyle(
-    val textContent: String = "",
-    val fontFamily: String = "Arial",
+    val textContent: String = "New Text",
+    val fontFamily: String = "LiberationSans",
     val fontSize: Float = 14f,
     val fontWeight: UnityFontWeight = UnityFontWeight.NORMAL,
     val textColor: String = "#000000",
     val textAlign: UnityTextAlign = UnityTextAlign.LEFT,
-    val lineSpacing: Float = 18f,
-    val letterSpacing: Float = 0f
 )
